Extract a few things. Added bearerToken method to request.
